+++ packages/graphico/src/types.ts
@@ -94,3 +94,7 @@ export interface Drawable {
      */
     draw(graphics: CanvasRenderingContext2D): void;
 }
+/**
+ * An arbitrary object which can be saved or loaded from browser storage.
+ */
+export type StoreData = Record<string, unknown>;